Change the data output and method of injection.
We now build the data used by the chrome extension in
OctoExtractor.build_extension_data which formats it in a way that we can
easily insert based on the current path of the url.

Now we only inject once per method so links dont get duplicated on the docs
page.